PlayerManager initial checkin
