+++ Source/ACE.Server/Network/Handlers/CharacterHandler.cs
@@ -148,7 +148,7 @@ namespace ACE.Server.Network.Handlers
                         return;
                     }
 
-                    WorldManager.AddPlayer(player.Character);
+                    PlayerManager.AddPlayer(player.Character);
                     session.Characters.Add(player.Character);
 
                     SendCharacterCreateResponse(session, CharacterGenerationVerificationResponse.Ok, player.Guid, characterCreateInfo.Name);
