+++ Source/ACE.Server/WorldObjects/Player_Allegiance.cs
@@ -32,10 +32,10 @@ namespace ACE.Server.WorldObjects
 
             Patron = targetGuid.Full;
             Monarch = AllegianceManager.GetMonarch(patron).Guid.Full;
-            WorldManager.SyncOffline(this);
+            PlayerManager.SyncOffline(this);
 
-            Console.WriteLine("Patron: " + WorldManager.GetOfflinePlayerByGuidId(Patron.Value).Name);
-            Console.WriteLine("Monarch: " + WorldManager.GetOfflinePlayerByGuidId(Monarch.Value).Name);
+            Console.WriteLine("Patron: " + PlayerManager.GetOfflinePlayerByGuidId(Patron.Value).Name);
+            Console.WriteLine("Monarch: " + PlayerManager.GetOfflinePlayerByGuidId(Monarch.Value).Name);
 
             // send message to patron:
             // %vassal% has sworn Allegiance to you.
@@ -63,7 +63,7 @@ namespace ACE.Server.WorldObjects
         {
             if (!IsBreakable(targetGuid)) return;
 
-            var target = WorldManager.GetOfflinePlayerByGuidId(targetGuid.Full);
+            var target = PlayerManager.GetOfflinePlayerByGuidId(targetGuid.Full);
 
             Console.WriteLine(Name + " breaking allegiance to " + target.Name);
 
@@ -76,13 +76,13 @@ namespace ACE.Server.WorldObjects
             {
                 target.Patron = null;
                 target.Monarch = null;
-                WorldManager.SyncOffline(target);
+                PlayerManager.SyncOffline(target);
             }
             else
             {
                 Patron = null;
                 Monarch = null;
-                WorldManager.SyncOffline(this);
+                PlayerManager.SyncOffline(this);
             }
 
             // send message to target if online
@@ -116,7 +116,7 @@ namespace ACE.Server.WorldObjects
             // player already sworn?
             if (Patron != null)
             {
-                Console.WriteLine(Name + " tried to swear to " + target.Name + ", but is already sworn to " + WorldManager.GetOfflinePlayerByGuidId(Patron.Value).Name);
+                Console.WriteLine(Name + " tried to swear to " + target.Name + ", but is already sworn to " + PlayerManager.GetOfflinePlayerByGuidId(Patron.Value).Name);
                 return false;
             }
 
@@ -172,7 +172,7 @@ namespace ACE.Server.WorldObjects
             // players can break from either vassals or patrons
 
             // ensure target player exists
-            var target = WorldManager.GetOfflinePlayerByGuidId(targetGuid.Full);
+            var target = PlayerManager.GetOfflinePlayerByGuidId(targetGuid.Full);
             if (target == null)
             {
                 Console.WriteLine(Name + " tried to break allegiance to an unknown player guid: " + targetGuid.Full.ToString("X8"));
